can prepare image for upload
